#2320 code review fix
