+++ kafka-ui-react-app/src/components/Connect/Details/Tasks/TaskRow.styled.ts
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const Cell = styled.td`
+  vertical-align: top !important;
+`;
