--- kafka-ui-react-app/src/components/Connect/Details/Tasks/TraceCell.styled.ts
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-export const Wrapper = styled.div`
-  display: flex;
-`;
-
-export const ToggleButton = styled.div`
-  padding: 4px 8px 0px 0px !important;
-`;
